+++ src/raft/RaftDirector.cc
@@ -26,7 +26,8 @@
 #include "RaftReplicator.hh"
 #include "RaftLease.hh"
 #include "../Dispatcher.hh"
-using namespace quarkdb;
+
+namespace quarkdb {
 
 RaftDirector::RaftDirector(RaftJournal &jour, StateMachine &sm, RaftState &st, RaftLease &ls, RaftCommitTracker &ct, RaftClock &rc, RaftWriteTracker &wt, ShardDirectory &sharddir, RaftConfig &conf, RaftReplicator &rep, const RaftContactDetails &cd)
 : journal(jour), stateMachine(sm), state(st), raftClock(rc), lease(ls), commitTracker(ct), writeTracker(wt), shardDirectory(sharddir), config(conf), replicator(rep), contactDetails(cd) {
@@ -82,6 +83,10 @@ void RaftDirector::leaderLoop(RaftStateSnapshotPtr &snapshot) {
 }
 
 void RaftDirector::runForLeader() {
+  // If we get vetoed, this ensures we stop election attempts up until the
+  // point we receive a fresh heartbeat.
+  std::chrono::steady_clock::time_point lastHeartbeat = raftClock.getLastHeartbeat();
+
   // don't reuse the snapshot from the main loop,
   // it could have changed in-between
   RaftStateSnapshotPtr snapshot = state.getSnapshot();
@@ -100,8 +105,15 @@ void RaftDirector::runForLeader() {
     return;
   }
 
-  if(RaftElection::perform(votereq, state, lease, contactDetails) != ElectionOutcome::kElected) {
+  ElectionOutcome electionOutcome = RaftElection::perform(votereq, state, lease, contactDetails);
+
+  if(electionOutcome != ElectionOutcome::kElected) {
     state.dropOut(snapshot->term+1);
+  }
+
+  if(electionOutcome == ElectionOutcome::kVetoed) {
+    lastHeartbeatBeforeVeto = lastHeartbeat;
+    qdb_info("Election round for term " << snapshot->term + 1 << " resulted in a veto. This means, the next leader of this cluster cannot be me. Stopping election attempts until I receive a heartbeat.");
   }
 }
 
@@ -114,7 +126,22 @@ void RaftDirector::followerLoop(RaftStateSnapshotPtr &snapshot) {
 
     writeTracker.flushQueues(Formatter::err("unavailable"));
     state.wait(randomTimeout);
-    if(raftClock.timeout()) {
+
+    if(raftClock.getLastHeartbeat() == lastHeartbeatBeforeVeto) {
+      // I've been vetoed during my last election attempt, and no heartbeat has
+      // appeared since then.
+      //
+      // It could be a network connectivity issue, where I'm able to establish
+      // TCP connections to other nodes (and thus disrupt them), but they cannot
+      // send me heartbeats.
+      //
+      // It could also be that I'm not a full member of this cluster, but I
+      // don't know it yet, and I'm being disruptive to the other nodes.
+      //
+      // Since a veto means the next cluster leader cannot be me, completely
+      // abstain from starting elections until we receive a heartbeat.
+    }
+    else if(raftClock.timeout()) {
       if(contains(journal.getMembership().nodes, state.getMyself())) {
         qdb_event(state.getMyself().toString() <<  ": TIMEOUT after " << randomTimeout.count() << "ms, I am not receiving heartbeats. Attempting to start election.");
         runForLeader();
@@ -123,4 +150,6 @@ void RaftDirector::followerLoop(RaftStateSnapshotPtr &snapshot) {
       qdb_warn("I am not receiving heartbeats - not running for leader since in membership epoch " << journal.getEpoch() << " I am not a full node. Will keep on waiting.");
     }
   }
+}
+
 }